requirements.txt
colorama==0.4.6
decorator==5.2.1
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
moviepy==2.2.1
numpy==2.4.0
pillow==11.3.0
proglog==0.1.12
python-dotenv==1.2.1
tqdm==4.67.1
